+++ backend/src/main/java/com/icehufs/icebreaker/domain/membership/dto/response/PatchUserPassResponseDto.java
@@ -1,4 +1,4 @@
-package com.icehufs.icebreaker.dto.response.user;
+package com.icehufs.icebreaker.domain.membership.dto.response;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -7,19 +7,19 @@ import com.icehufs.icebreaker.common.ResponseCode;
 import com.icehufs.icebreaker.common.ResponseMessage;
 import com.icehufs.icebreaker.dto.response.ResponseDto;
 
-public class DeleteUserResponseDto extends ResponseDto{
-    private DeleteUserResponseDto(){
+public class PatchUserPassResponseDto extends ResponseDto{
+
+        private PatchUserPassResponseDto(){
         super(ResponseCode.SUCCESS, ResponseMessage.SUCCESS);
     }
 
-        public static ResponseEntity<DeleteUserResponseDto> success(){
-            DeleteUserResponseDto result = new DeleteUserResponseDto();
+        public static ResponseEntity<PatchUserPassResponseDto> success(){
+            PatchUserPassResponseDto result = new PatchUserPassResponseDto();
         return ResponseEntity.status(HttpStatus.OK).body(result);
     }
         public static ResponseEntity<ResponseDto> notExistUser (){
         ResponseDto result = new ResponseDto(ResponseCode.NOT_EXISTED_USER, ResponseMessage.NOT_EXISTED_USER);
         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(result);
     }
-
     
 }
